+++ packages/mongo_document_db/lib/src/database/commands/base/command_operation.dart
@@ -101,7 +101,9 @@ class CommandOperation extends OperationBase {
     var modernMessage = MongoModernMessage(command);
 
     return db.executeModernMessage(modernMessage,
-        connection: connection, skipStateCheck: skipStateCheck);
+        connection: connection,
+        skipStateCheck: skipStateCheck,
+        replayReadsUntilSelectionTimeout: !hasAspect(Aspect.writeOperation));
   }
 }